backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.0
Flask-Login==0.5.0
PyJWT==2.1.0
Werkzeug==2.0.1
SQLAlchemy<2.0
Flask-SQLAlchemy
Flask-Login
